Add JavaDoc comment for test files
