+++ src/test/java/seedu/address/logic/commands/stubs/FindPersonArgumentParseResultStub.java
@@ -5,9 +5,16 @@ import java.util.List;
 
 import seedu.address.logic.parser.commandoption.Option;
 
+/**
+ * Mock stub for ArgumentParseResult to test FindCommand. Only stores the
+ * list of keywords to find
+ */
 public class FindPersonArgumentParseResultStub extends ArgumentParseResultStub {
     private List<String> keywords;
 
+    /**
+     * @param keywords The string of keywords to find, internally parses into a List
+     */
     public FindPersonArgumentParseResultStub(String keywords) {
         super();
         this.keywords = Arrays.asList(keywords.split("\\s+"));
