--- src/test/java/seedu/address/logic/commands/stubs/ModelStubWithPerson.java
@@ -1,30 +1,0 @@
-package seedu.address.logic.commands.stubs;
-
-import static java.util.Objects.requireNonNull;
-
-import seedu.address.model.person.Person;
-
-/**
- * A Model stub that contains a single person.
- */
-public class ModelStubWithPerson extends ModelStub {
-    private Person person;
-
-    public ModelStubWithPerson(Person person) {
-        requireNonNull(person);
-        this.person = person;
-    }
-
-    @Override
-    public boolean hasPerson(Person person) {
-        requireNonNull(person);
-        return this.person.isSamePerson(person);
-    }
-
-    @Override
-    public void deletePerson(Person target) {
-        if (target.equals(person)) {
-            target = null;
-        }
-    }
-}
